refactor: split out function definer and tests into separate files
