+++ src/definers/functions/sync-tag-invalidate.ts
@@ -1,0 +1,51 @@
+import {
+  type BlueprintSyncTagInvalidateFunctionConfig,
+  type BlueprintSyncTagInvalidateFunctionResource,
+  validateSyncTagInvalidateFunction,
+} from '../../index.js'
+import { runValidation } from '../../utils/validation.js'
+import { defineFunction } from './index.js'
+/**
+ * Defines a function that is triggered on a sync tag invalidate event.
+ *
+ * @remarks
+ * Using implicit event scoping to all datasets in the function project:
+ * ```ts
+ * defineSyncTagInvalidateFunction({
+ *   name: 'bustin-caches',
+ * })
+ * ```
+ *
+ * Using explicit event scoping to a particular dataset in the function project:
+ * ```ts
+ * defineSyncTagInvalidateFunction({
+ *   name: 'bustin-caches',
+ *   event: {
+ *     resource: {
+ *       type: 'dataset',
+ *       id: 'myProj.myDataset',
+ *     },
+ *   },
+ * })
+ * ```
+ * @public
+ * @category Definers
+ * @expandType BlueprintSyncTagInvalidateFunctionConfig
+ * @param functionConfig The configuration for the sync tag invalidate function
+ * @returns The validated sync tag invalidate function resource
+ */
+export function defineSyncTagInvalidateFunction(
+  functionConfig: BlueprintSyncTagInvalidateFunctionConfig,
+): BlueprintSyncTagInvalidateFunctionResource {
+  const { event } = functionConfig
+
+  const functionResource: BlueprintSyncTagInvalidateFunctionResource = {
+    ...defineFunction(functionConfig, { skipValidation: true }),
+    type: 'sanity.function.sync-tag-invalidate',
+    ...(event?.resource ? { event } : {}),
+  }
+
+  runValidation(() => validateSyncTagInvalidateFunction(functionResource))
+
+  return functionResource
+}
