+++ src/index.ts
@@ -10,7 +10,14 @@
 export * from './definers/blueprints.js'
 export * from './definers/cors.js'
 export * from './definers/datasets.js'
-export * from './definers/functions.js'
+export * from './definers/functions/document.js'
+export * from './definers/functions/event.js'
+export * from './definers/functions/index.js'
+export * from './definers/functions/media-library.js'
+export * from './definers/functions/pipeline.js'
+export * from './definers/functions/queue.js'
+export * from './definers/functions/scheduled.js'
+export * from './definers/functions/sync-tag-invalidate.js'
 export * from './definers/projects.js'
 export * from './definers/resources.js'
 export * from './definers/robots.js'
