--- test/unit/definers/functions.test.ts
@@ -1,434 +1,0 @@
-import {afterEach, describe, expect, test, vi} from 'vitest'
-import * as fns from '../../../src/definers/functions.js'
-import type {BlueprintFunctionResourceEvent} from '../../../src/index.js'
-import * as index from '../../../src/index.js'
-import {defineBlueprintForResource} from '../../helpers/index.js'
-
-vi.mock(import('../../../src/index.js'), async (importOriginal) => {
-  const originalModule = await importOriginal()
-  return {
-    ...originalModule,
-    validateBlueprint: vi.fn(() => []),
-  }
-})
-
-describe('defineFunction', () => {
-  describe('happy paths', () => {
-    test('should assign src based on name if not provided', () => {
-      const fn = fns.defineFunction({name: 'test'})
-      expect(fn.src).toEqual('functions/test')
-    })
-
-    test('should ignore invalid properties', () => {
-      // @ts-expect-error Intentionally wrong type
-      const fn = fns.defineFunction({name: 'test', invalid: 'invalid'})
-      expect(Object.keys(fn)).not.toContain('invalid')
-    })
-  })
-  describe('sad paths', () => {
-    afterEach(() => {
-      vi.resetAllMocks()
-    })
-
-    test('should throw an error if validateFunction returns an error', () => {
-      const spy = vi.spyOn(index, 'validateFunction').mockImplementation(() => [{type: 'test', message: 'this is a test'}])
-      expect(() => defineBlueprintForResource(fns.defineFunction({name: 'func-name'}))).toThrow('this is a test')
-
-      expect(spy).toHaveBeenCalledOnce()
-    })
-  })
-})
-describe('defineDocumentFunction', () => {
-  describe('happy paths', () => {
-    test('should create a default publish event with provided filter', () => {
-      const fn = fns.defineDocumentFunction({name: 'test', event: {filter: '_type == "post"'}})
-      expect(fn.event).toEqual({on: ['publish'], filter: '_type == "post"'})
-    })
-
-    test('should throw an error if event keys are defined using a mix of under the event object as well as at the top level', () => {
-      expect(() =>
-        fns.defineDocumentFunction({
-          name: 'test',
-          event: {on: ['publish']},
-          filter: '_type == "post"',
-        }),
-      ).toThrowError(/`event` properties should be specified under the `event` key/i)
-    })
-
-    test('should create the event with publish if not provided', () => {
-      const fn = fns.defineDocumentFunction({name: 'test', src: 'test.js'})
-      expect(fn.event).toEqual({on: ['publish']})
-    })
-
-    test('should allow for creating events triggered on create, update and delete', () => {
-      const fn = fns.defineDocumentFunction({name: 'test', src: 'test.js', event: {on: ['create', 'update', 'delete']}})
-      expect(fn.event.on).toEqual(['create', 'update', 'delete'])
-    })
-
-    test('should allow for creating events with explicit include* toggles', () => {
-      const fn = fns.defineDocumentFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update'], includeAllVersions: true, includeDrafts: true},
-      })
-      expect(fn.event.includeDrafts).toEqual(true)
-      expect(fn.event.includeAllVersions).toEqual(true)
-    })
-
-    test('should allow for creating events scoped to a specific dataset', () => {
-      const fn = fns.defineDocumentFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update'], resource: {type: 'dataset', id: 'myProject.myDataset'}},
-      })
-      expect(fn.event.resource?.type).toEqual('dataset')
-      expect(fn.event.resource?.id).toEqual('myProject.myDataset')
-    })
-
-    test('should allow for creating events explicitly scoped to all datasets', () => {
-      const fn = fns.defineDocumentFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update'], resource: {type: 'dataset', id: 'myProject.*'}},
-      })
-      expect(fn.event.resource?.type).toEqual('dataset')
-      expect(fn.event.resource?.id).toEqual('myProject.*')
-    })
-
-    test('should allow for creating a function in a specific project', () => {
-      const fn = fns.defineDocumentFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update']},
-        project: 'projectId',
-      })
-      expect(fn.project).toEqual('projectId')
-    })
-  })
-
-  describe('sad paths', () => {
-    afterEach(() => {
-      vi.resetAllMocks()
-    })
-
-    test('should throw an error if validateDocumentFunction returns an error', () => {
-      const spy = vi.spyOn(index, 'validateDocumentFunction').mockImplementation(() => [{type: 'test', message: 'this is a test'}])
-      expect(() => defineBlueprintForResource(fns.defineDocumentFunction({name: 'test', event: {on: ['publish']}}))).toThrow(
-        'this is a test',
-      )
-
-      expect(spy).toHaveBeenCalledOnce()
-    })
-  })
-})
-
-describe('defineMediaLibraryAssetFunction', () => {
-  const resource = {type: 'media-library' as const, id: 'ml12345'}
-  describe('happy paths', () => {
-    test('should create a default publish event with provided filter', () => {
-      const event = {filter: '_type == "post"', resource}
-      const fn = fns.defineMediaLibraryAssetFunction({
-        name: 'test',
-        event,
-      })
-      expect(fn.event.on).toEqual(['publish'])
-      expect(fn.event.includeDrafts).toBeFalsy()
-    })
-
-    test('should create the event with publish if not provided', () => {
-      const fn = fns.defineMediaLibraryAssetFunction({name: 'test', src: 'test.js', event: {resource}})
-      expect(fn.event.on).toEqual(['publish'])
-    })
-
-    test('should allow for creating events triggered on create, update and delete', () => {
-      const fn = fns.defineMediaLibraryAssetFunction({name: 'test', src: 'test.js', event: {on: ['create', 'update', 'delete'], resource}})
-      expect(fn.event.on).toEqual(['create', 'update', 'delete'])
-    })
-
-    test('should allow for creating events with explicit include* toggles', () => {
-      let fn = fns.defineMediaLibraryAssetFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update'], includeDrafts: true, resource},
-      })
-      expect(fn.event.includeDrafts).toEqual(true)
-      fn = fns.defineMediaLibraryAssetFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update'], includeDrafts: false, resource},
-      })
-      expect(fn.event.includeDrafts).toEqual(false)
-    })
-
-    test('should allow for creating functions in a specific project', () => {
-      const fn = fns.defineMediaLibraryAssetFunction({
-        name: 'test',
-        src: 'test.js',
-        event: {on: ['update'], resource},
-        project: 'projectId',
-      })
-      expect(fn.project).toEqual('projectId')
-    })
-  })
-
-  describe('sad paths', () => {
-    afterEach(() => {
-      vi.resetAllMocks()
-    })
-
-    test('should throw an error if validateMediaLibraryAssetFunction returns an error', () => {
-      const spy = vi.spyOn(index, 'validateMediaLibraryAssetFunction').mockImplementation(() => [{type: 'test', message: 'this is a test'}])
-      expect(() =>
-        defineBlueprintForResource(
-          fns.defineMediaLibraryAssetFunction({name: 'test', event: {on: ['publish'], resource: {type: 'media-library', id: 'ml1234'}}}),
-        ),
-      ).toThrow('this is a test')
-
-      expect(spy).toHaveBeenCalledOnce()
-    })
-  })
-})
-
-describe('defineScheduledFunction', () => {
-  describe('happy paths', () => {
-    test('should create a scheduled event with explicit cron fields', () => {
-      const event = {minute: '*', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'}
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event,
-      })
-      expect(fn.event).toEqual({minute: '*', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'})
-    })
-
-    test('should normalize expression to explicit cron fields', () => {
-      const event = {expression: '* * * * *'}
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event,
-      })
-      expect(fn.event).toEqual({minute: '*', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'})
-    })
-
-    test('should create a scheduled function with optional timezone', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: '* * * * *'},
-        timezone: 'America/New_York',
-      })
-      expect(fn.timezone).toEqual('America/New_York')
-      expect(fn.event).toEqual({minute: '*', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'})
-    })
-
-    test('should parse natural language expression to explicit cron fields', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: 'every day at 9am'},
-      })
-      expect(fn.event).toEqual({minute: '0', hour: '9', dayOfMonth: '*', month: '*', dayOfWeek: '*'})
-    })
-
-    test('should normalize cron expression to explicit fields', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: '0 9 * * 1-5'},
-      })
-      expect(fn.event).toEqual({minute: '0', hour: '9', dayOfMonth: '*', month: '*', dayOfWeek: '1-5'})
-    })
-
-    test('should parse weekday natural language', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: 'weekdays at 8am'},
-      })
-      expect(fn.event).toEqual({minute: '0', hour: '8', dayOfMonth: '*', month: '*', dayOfWeek: '1-5'})
-    })
-
-    test('should parse specific weekday with time', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: 'fridays at 9am'},
-      })
-      expect(fn.event).toEqual({minute: '0', hour: '9', dayOfMonth: '*', month: '*', dayOfWeek: '5'})
-    })
-
-    test('should parse time of day period', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: 'fridays in the evening'},
-      })
-      expect(fn.event).toEqual({minute: '0', hour: '18', dayOfMonth: '*', month: '*', dayOfWeek: '5'})
-    })
-
-    test('should parse multiple weekdays', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: 'mon, wed, fri at 9am'},
-      })
-      expect(fn.event).toEqual({minute: '0', hour: '9', dayOfMonth: '*', month: '*', dayOfWeek: '1,3,5'})
-    })
-
-    test('should parse interval schedules', () => {
-      const fn = fns.defineScheduledFunction({
-        name: 'test',
-        event: {expression: 'every 15 minutes'},
-      })
-      expect(fn.event).toEqual({minute: '*/15', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'})
-    })
-
-    test('deprecated method defineScheduleFunction should work', () => {
-      const fn = fns.defineScheduleFunction({
-        name: 'test',
-        event: {expression: 'every 15 minutes'},
-      })
-      expect(fn.event).toEqual({minute: '*/15', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'})
-    })
-  })
-
-  describe('sad paths', () => {
-    afterEach(() => {
-      vi.resetAllMocks()
-    })
-
-    test('should throw an error if validateScheduledFunction returns an error', () => {
-      const spy = vi.spyOn(index, 'validateScheduledFunction').mockImplementation(() => [{type: 'test', message: 'this is a test'}])
-      expect(() =>
-        defineBlueprintForResource(
-          fns.defineScheduledFunction({name: 'test', event: {minute: '*', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'}}),
-        ),
-      ).toThrow('this is a test')
-
-      expect(spy).toHaveBeenCalledOnce()
-    })
-  })
-})
-
-describe('definePipeline', () => {
-  describe('happy paths', () => {
-    test('should create a pipeline event', () => {
-      const fn = fns.definePipeline({
-        name: 'test',
-      })
-      expect(fn.name).toEqual('test')
-    })
-
-    test('should have the default src path', () => {
-      const fn = fns.definePipeline({
-        name: 'test',
-      })
-      expect(fn.src).toEqual('functions/test')
-    })
-
-    test('should create a pipeline with an event', () => {
-      const fn = fns.definePipeline({
-        name: 'test',
-        event: {type: 'document', on: ['create'], filter: "_type == 'article'"},
-      })
-      expect(fn.event).toEqual({type: 'document', on: ['create'], filter: "_type == 'article'"})
-    })
-
-    test('should create a pipeline function with optional concurrency', () => {
-      const fn = fns.definePipeline({
-        name: 'test',
-        concurrency: 3,
-      })
-      expect(fn.concurrency).toEqual(3)
-      expect(fn.name).toEqual('test')
-    })
-
-    test('should create a pipeline function with optional debounce', () => {
-      const fn = fns.definePipeline({
-        name: 'test',
-        debounce: 3,
-      })
-      expect(fn.debounce).toEqual(3)
-      expect(fn.name).toEqual('test')
-    })
-
-    test('should create a pipeline function with optional debounceKey', () => {
-      const fn = fns.definePipeline({
-        name: 'test',
-        debounce: 1,
-        debounceKey: 'testKey',
-      })
-      expect(fn.debounce).toEqual(1)
-      expect(fn.debounceKey).toEqual('testKey')
-      expect(fn.name).toEqual('test')
-    })
-
-    describe('sad paths', () => {
-      afterEach(() => {
-        vi.resetAllMocks()
-      })
-
-      test('should throw an error if validatePipelineFunction returns an error', () => {
-        const spy = vi.spyOn(index, 'validatePipelineFunction').mockImplementation(() => [{type: 'test', message: 'this is a test'}])
-        expect(() =>
-          defineBlueprintForResource(
-            fns.definePipeline({name: 'test', event: {type: 'document', on: ['create'], filter: "_type == 'article'"}}),
-          ),
-        ).toThrow('this is a test')
-
-        expect(spy).toHaveBeenCalledOnce()
-      })
-    })
-  })
-
-  describe('defineQueueFunction', () => {
-    describe('happy paths', () => {
-      test('should create a queue function without an event', () => {
-        const fn = fns.defineQueueFunction({name: 'test'})
-        expect(fn.type).toEqual('sanity.function.queue')
-        expect(fn).not.toHaveProperty('event')
-      })
-
-      test('should pass through a document event', () => {
-        const event: BlueprintFunctionResourceEvent = {type: 'document', on: ['publish'], filter: "_type == 'post'"}
-        const fn = fns.defineQueueFunction({name: 'test', event})
-        expect(fn.event).toEqual(event)
-      })
-
-      test('should pass through a media-library event', () => {
-        const event: BlueprintFunctionResourceEvent = {
-          type: 'media-library',
-          on: ['create'],
-          resource: {type: 'media-library', id: 'my-media-library-id'},
-        }
-        const fn = fns.defineQueueFunction({name: 'test', event})
-        expect(fn.event).toEqual(event)
-      })
-
-      test('should pass through a cron event', () => {
-        const event: BlueprintFunctionResourceEvent = {type: 'cron', minute: '*', hour: '*', dayOfMonth: '*', month: '*', dayOfWeek: '*'}
-        const fn = fns.defineQueueFunction({name: 'test', event})
-        expect(fn.event).toEqual(event)
-      })
-
-      test('should pass through a sync-tag-invalidate event', () => {
-        const event: BlueprintFunctionResourceEvent = {type: 'sync-tag-invalidate', resource: {type: 'dataset', id: 'myProj.myDataset'}}
-        const fn = fns.defineQueueFunction({name: 'test', event})
-        expect(fn.event).toEqual(event)
-      })
-
-      test('should pass through concurrency, fifo and dlq alongside an event', () => {
-        const event: BlueprintFunctionResourceEvent = {type: 'document', on: ['publish']}
-        const fn = fns.defineQueueFunction({name: 'test', event, concurrency: 5, fifo: false, dlq: false})
-        expect(fn).toMatchObject({event, concurrency: 5, fifo: false, dlq: false})
-      })
-    })
-
-    describe('sad paths', () => {
-      afterEach(() => {
-        vi.resetAllMocks()
-      })
-
-      test('should throw an error if validateQueueFunction returns an error', () => {
-        const spy = vi.spyOn(index, 'validateQueueFunction').mockImplementation(() => [{type: 'test', message: 'this is a test'}])
-        expect(() =>
-          defineBlueprintForResource(fns.defineQueueFunction({name: 'test', event: {type: 'document', on: ['publish']}})),
-        ).toThrow('this is a test')
-
-        expect(spy).toHaveBeenCalledOnce()
-      })
-    })
-  })
-})
